+++ source/voxelman/geometry/box.d
@@ -1,107 +1,28 @@
 /**
-Copyright: Copyright (c) 2015-2016 Andrey Penechko.
+Copyright: Copyright (c) 2016 Andrey Penechko.
 License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
 Authors: Andrey Penechko.
 */
-module voxelman.world.storage.volume;
+module voxelman.geometry.box;
 
-import std.experimental.logger;
 import std.algorithm : alg_min = min, alg_max = max;
-import std.math : floor;
 import std.range : chain, only;
 import dlib.math.vector;
 
-import voxelman.core.config;
-import voxelman.world.storage.coordinates;
-
-import voxelman.utils.renderutils;
-void putCube(ref Batch batch, Volume volume, Color3ub color, bool fill, bool offset = true)
+Box boxFromCorners(ivec3 a, ivec3 b)
 {
-	vec3 pos = volume.position;
-	vec3 size = volume.size;
-	if (offset) {
-		pos -= vec3(0.01, 0.01, 0.01);
-		size += vec3(0.02, 0.02, 0.02);
-	}
-	batch.putCube(pos, size, color, fill);
+	Box box;
+	box.position = min(a, b);
+	box.size = max(a, b) - box.position + ivec3(1,1,1);
+	return box;
 }
 
-Volume calcVolume(ChunkWorldPos cwp, int viewRadius)
-{
-	int size = viewRadius*2 + 1;
-	return Volume(cast(ivec3)(cwp.ivector3 - viewRadius),
-		ivec3(size, size, size), cwp.w);
-}
-
-Volume volumeFromCorners(ivec3 a, ivec3 b, DimentionId dimention = 0)
-{
-	Volume vol;
-	vol.position = min(a, b);
-	vol.size = max(a, b) - vol.position + ivec3(1,1,1);
-	vol.dimention = dimention;
-	return vol;
-}
-
-Volume blockVolumeToChunkVolume(Volume blockVolume)
-{
-	auto startPosition = blockToChunkPosition(blockVolume.position);
-	auto endPosition = blockToChunkPosition(blockVolume.endPosition);
-	return volumeFromCorners(startPosition, endPosition, blockVolume.dimention);
-}
-
-// makes block volume in chunk-local space out of world space
-Volume blockVolumeToChunkLocalVolume(Volume blockVolume)
-{
-	blockVolume.position -= chunkStartBlockPos(blockVolume.position);
-	return blockVolume;
-}
-
-/// Returns chunks if their mesh may have changed after specified modification
-/// Volume is specified in block space
-Volume calcModifiedMeshesVolume(Volume modificationVolume)
-{
-	// We increase size by 1 in every direction.
-	// After rounding chunks on the border of modification will be included
-	Volume expandedVolume = modificationVolume;
-	expandedVolume.position -= ivec3(1,1,1);
-	expandedVolume.size += ivec3(2,2,2);
-	Volume chunkVolume = blockVolumeToChunkVolume(expandedVolume);
-	return chunkVolume;
-}
-
-Volume chunkToBlockVolume(ChunkWorldPos cwp) {
-	return chunkToBlockVolume(cwp.ivector3, cwp.dimention);
-}
-
-Volume chunkToBlockVolume(ivec3 cwp, ushort dimention) {
-	ivec3 startPosition = chunkToBlockPosition(cwp);
-	return Volume(startPosition, CHUNK_SIZE_VECTOR, dimention);
-}
-
-Vector!(T, size) min(T, int size)(Vector!(T, size) a, Vector!(T, size) b)
-{
-	Vector!(T, size) res;
-	foreach(i; 0..size)
-		res.arrayof[i] = alg_min(a.arrayof[i], b.arrayof[i]);
-	return res;
-}
-
-Vector!(T, size) max(T, int size)(Vector!(T, size) a, Vector!(T, size) b)
-{
-	Vector!(T, size) res;
-	foreach(i; 0..size)
-		res.arrayof[i] = alg_max(a.arrayof[i], b.arrayof[i]);
-	return res;
-}
-
-// 3d grid volume
-struct Volume
+struct Box
 {
 	ivec3 position;
 	ivec3 size;
-	ushort dimention;
 
-	int volume() @property const
+	int box() @property const
 	{
 		return size.x * size.y * size.z;
 	}
@@ -124,17 +45,11 @@ struct Volume
 		return true;
 	}
 
-	bool contains(ivec3 point, ushort dimention) const
-	{
-		if (this.dimention != dimention) return false;
-		return contains(point);
-	}
-
 	import std.algorithm : cartesianProduct, map, joiner, equal, canFind;
 	import std.range : iota, walkLength;
 	import std.array : array;
 
-	// generates all positions within volume.
+	// generates all positions within box.
 	auto positions() const @property
 	{
 		return cartesianProduct(
@@ -146,23 +61,23 @@ struct Volume
 
 	unittest
 	{
-		assert(Volume(ivec3(0,0,0), ivec3(3,3,3)).positions.walkLength == 27);
+		assert(Box(ivec3(0,0,0), ivec3(3,3,3)).positions.walkLength == 27);
 	}
 }
 
 struct TrisectResult
 {
-	Volume[] aVolumes;
-	Volume intersection;
-	Volume[] bVolumes;
+	Box[] aBoxes;
+	Box intersection;
+	Box[] bBoxes;
 	import std.algorithm : map, joiner;
 	auto aPositions() @property
 	{
-		return aVolumes.map!(a => a.positions).joiner;
+		return aBoxes.map!(a => a.positions).joiner;
 	}
 	auto bPositions() @property
 	{
-		return bVolumes.map!(a => a.positions).joiner;
+		return bBoxes.map!(a => a.positions).joiner;
 	}
 }
 
@@ -170,15 +85,15 @@ struct TrisectResult
 // TrisectResult[0] a - b  == a if no intersection
 // TrisectResult[1] a intersects b
 // TrisectResult[2] b - a  == b if no intersection
-TrisectResult trisect(Volume a, Volume b)
+TrisectResult trisect(Box a, Box b)
 {
 	TrisectResult result;
-	result.intersection = volumeIntersection(a, b);
+	result.intersection = boxIntersection(a, b);
 
 	// no intersection
 	if (result.intersection.empty)
 	{
-		return TrisectResult([a], Volume(), [b]);
+		return TrisectResult([a], Box(), [b]);
 	}
 
 	auto xTrisect = trisectAxis(a.position.x, a.position.x + a.size.x, b.position.x, b.position.x + b.size.x);
@@ -190,7 +105,7 @@ TrisectResult trisect(Volume a, Volume b)
 	foreach(za; zTrisect.aranges[0..zTrisect.numRangesA].chain(only(zTrisect.irange)))
 	{
 		if (!(xa.isIntersection && ya.isIntersection && za.isIntersection))
-			result.aVolumes ~= Volume(ivec3(xa.start, ya.start, za.start),
+			result.aBoxes ~= Box(ivec3(xa.start, ya.start, za.start),
 				ivec3(xa.length, ya.length, za.length));
 	}
 
@@ -199,7 +114,7 @@ TrisectResult trisect(Volume a, Volume b)
 	foreach(zb; zTrisect.branges[0..zTrisect.numRangesB].chain(only(zTrisect.irange)))
 	{
 		if (!(xb.isIntersection && yb.isIntersection && zb.isIntersection))
-			result.bVolumes ~= Volume(ivec3(xb.start, yb.start, zb.start),
+			result.bBoxes ~= Box(ivec3(xb.start, yb.start, zb.start),
 				ivec3(xb.length, yb.length, zb.length));
 	}
 
@@ -284,25 +199,19 @@ TrisectAxisResult trisectAxis(int aStart, int aEnd, int bStart, int bEnd)
 	return res;
 }
 
-Volume volumeIntersection(Volume a, Volume b)
+Box boxIntersection(Box a, Box b)
 {
-	if (a.dimention != b.dimention)
-	{
-		return Volume();
-	}
-
 	auto xTrisect = trisectAxis(a.position.x, a.position.x + a.size.x, b.position.x, b.position.x + b.size.x);
 	auto yTrisect = trisectAxis(a.position.y, a.position.y + a.size.y, b.position.y, b.position.y + b.size.y);
 	auto zTrisect = trisectAxis(a.position.z, a.position.z + a.size.z, b.position.z, b.position.z + b.size.z);
 
-	Volume result = Volume(
+	Box result = Box(
 			ivec3(xTrisect.irange.start, yTrisect.irange.start, zTrisect.irange.start),
-			ivec3(xTrisect.irange.length, yTrisect.irange.length, zTrisect.irange.length),
-			a.dimention);
+			ivec3(xTrisect.irange.length, yTrisect.irange.length, zTrisect.irange.length));
 
 	foreach(elem; result.size.arrayof) {
 		if (elem <= 0)
-			return Volume();
+			return Box();
 	}
 
 	return result;
@@ -310,28 +219,44 @@ Volume volumeIntersection(Volume a, Volume b)
 
 unittest
 {
-	assert(volumeIntersection(
-		Volume(ivec3(0,0,0), ivec3(2,2,2)),
-		Volume(ivec3(1,1,1), ivec3(2,2,2))) ==
-		Volume(ivec3(1,1,1), ivec3(1,1,1)));
-	assert(volumeIntersection(
-		Volume(ivec3(0,0,0), ivec3(2,2,2)),
-		Volume(ivec3(3,3,3), ivec3(4,4,4))) ==
-		Volume(ivec3()));
-	assert(volumeIntersection(
-		Volume(ivec3(1,1,1), ivec3(2,2,2)),
-		Volume(ivec3(0,0,0), ivec3(2,2,2))) ==
-		Volume(ivec3(1,1,1), ivec3(1,1,1)));
-	assert(volumeIntersection(
-		Volume(ivec3(1,1,1), ivec3(1,1,1)),
-		Volume(ivec3(1,1,1), ivec3(1,1,1))) ==
-		Volume(ivec3(1,1,1), ivec3(1,1,1)));
-	assert(volumeIntersection(
-		Volume(ivec3(0,0,0), ivec3(2,2,2)),
-		Volume(ivec3(0,0,-1), ivec3(2,2,2))) ==
-		Volume(ivec3(0,0,0), ivec3(2,2,1)));
-	assert(volumeIntersection(
-		Volume(ivec3(1,0,0), ivec3(1,1,1)),
-		Volume(ivec3(0,0,0), ivec3(32,32,32))) ==
-		Volume(ivec3(1,0,0), ivec3(1,1,1)));
+	assert(boxIntersection(
+		Box(ivec3(0,0,0), ivec3(2,2,2)),
+		Box(ivec3(1,1,1), ivec3(2,2,2))) ==
+		Box(ivec3(1,1,1), ivec3(1,1,1)));
+	assert(boxIntersection(
+		Box(ivec3(0,0,0), ivec3(2,2,2)),
+		Box(ivec3(3,3,3), ivec3(4,4,4))) ==
+		Box(ivec3()));
+	assert(boxIntersection(
+		Box(ivec3(1,1,1), ivec3(2,2,2)),
+		Box(ivec3(0,0,0), ivec3(2,2,2))) ==
+		Box(ivec3(1,1,1), ivec3(1,1,1)));
+	assert(boxIntersection(
+		Box(ivec3(1,1,1), ivec3(1,1,1)),
+		Box(ivec3(1,1,1), ivec3(1,1,1))) ==
+		Box(ivec3(1,1,1), ivec3(1,1,1)));
+	assert(boxIntersection(
+		Box(ivec3(0,0,0), ivec3(2,2,2)),
+		Box(ivec3(0,0,-1), ivec3(2,2,2))) ==
+		Box(ivec3(0,0,0), ivec3(2,2,1)));
+	assert(boxIntersection(
+		Box(ivec3(1,0,0), ivec3(1,1,1)),
+		Box(ivec3(0,0,0), ivec3(32,32,32))) ==
+		Box(ivec3(1,0,0), ivec3(1,1,1)));
+}
+
+Vector!(T, size) min(T, int size)(Vector!(T, size) a, Vector!(T, size) b)
+{
+	Vector!(T, size) res;
+	foreach(i; 0..size)
+		res.arrayof[i] = alg_min(a.arrayof[i], b.arrayof[i]);
+	return res;
+}
+
+Vector!(T, size) max(T, int size)(Vector!(T, size) a, Vector!(T, size) b)
+{
+	Vector!(T, size) res;
+	foreach(i; 0..size)
+		res.arrayof[i] = alg_max(a.arrayof[i], b.arrayof[i]);
+	return res;
 }